requirements.txt
selenium
requests
beautifulsoup4
psycopg2-binary
python-dotenv
